backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
rank-bm25==0.2.2
sentence-transformers==2.2.2
ollama==0.1.6
watchdog==3.0.0
pypdf==3.17.1
python-docx==1.1.0
langchain-text-splitters==0.0.1
